arch: consolidate all of the multiplexed syscall handling
Not only does this reduce the amount of duplicated code
significantly, it removes a lot of the "magic" numbers in the
code, and it happened to catch some bugs too.

Acked-by: Tom Hromatka <tom.hromatka@oracle.com>

(imported from commit 17cbd2c253ce63e5e9e3cec867ff58efbe8b5fdc)
