+++ src/arch-aarch64.c
@@ -31,8 +31,8 @@ const struct arch_def arch_def_aarch64 = {
 	.token_bpf = AUDIT_ARCH_AARCH64,
 	.size = ARCH_SIZE_64,
 	.endian = ARCH_ENDIAN_LITTLE,
-	.syscall_resolve_name = aarch64_syscall_resolve_name,
-	.syscall_resolve_num = aarch64_syscall_resolve_num,
+	.syscall_resolve_name_raw = aarch64_syscall_resolve_name,
+	.syscall_resolve_num_raw = aarch64_syscall_resolve_num,
 	.syscall_rewrite = NULL,
 	.rule_add = NULL,
 };
