+++ src/arch-mips64n32.c
@@ -32,6 +32,7 @@
 
 /**
  * Resolve a syscall name to a number
+ * @param arch the architecture definition
  * @param name the syscall name
  *
  * Resolve the given syscall name to the syscall number using the syscall table.
@@ -39,7 +40,8 @@
  * numbers; returns __NR_SCMP_ERROR on failure.
  *
  */
-int mips64n32_syscall_resolve_name_munge(const char *name)
+int mips64n32_syscall_resolve_name_munge(const struct arch_def *arch,
+					 const char *name)
 {
 	int sys;
 
@@ -53,6 +55,7 @@ int mips64n32_syscall_resolve_name_munge(const char *name)
 
 /**
  * Resolve a syscall number to a name
+ * @param arch the architecture definition
  * @param num the syscall number
  *
  * Resolve the given syscall number to the syscall name using the syscall table.
@@ -60,7 +63,8 @@ int mips64n32_syscall_resolve_name_munge(const char *name)
  * syscall names; returns NULL on failure.
  *
  */
-const char *mips64n32_syscall_resolve_num_munge(int num)
+const char *mips64n32_syscall_resolve_num_munge(const struct arch_def *arch,
+						int num)
 {
 	/* NOTE: we don't want to modify the pseudo-syscall numbers */
 	if (num >= __SCMP_NR_BASE)
@@ -74,7 +78,9 @@ const struct arch_def arch_def_mips64n32 = {
 	.size = ARCH_SIZE_32,
 	.endian = ARCH_ENDIAN_BIG,
 	.syscall_resolve_name = mips64n32_syscall_resolve_name_munge,
+	.syscall_resolve_name_raw = mips64n32_syscall_resolve_name,
 	.syscall_resolve_num = mips64n32_syscall_resolve_num_munge,
+	.syscall_resolve_num_raw = mips64n32_syscall_resolve_num,
 	.syscall_rewrite = NULL,
 	.rule_add = NULL,
 };
@@ -85,7 +91,9 @@ const struct arch_def arch_def_mipsel64n32 = {
 	.size = ARCH_SIZE_32,
 	.endian = ARCH_ENDIAN_LITTLE,
 	.syscall_resolve_name = mips64n32_syscall_resolve_name_munge,
+	.syscall_resolve_name_raw = mips64n32_syscall_resolve_name,
 	.syscall_resolve_num = mips64n32_syscall_resolve_num_munge,
+	.syscall_resolve_num_raw = mips64n32_syscall_resolve_num,
 	.syscall_rewrite = NULL,
 	.rule_add = NULL,
 };
