+++ src/arch-parisc.c
@@ -15,8 +15,8 @@ const struct arch_def arch_def_parisc = {
 	.token_bpf = AUDIT_ARCH_PARISC,
 	.size = ARCH_SIZE_32,
 	.endian = ARCH_ENDIAN_BIG,
-	.syscall_resolve_name = parisc_syscall_resolve_name,
-	.syscall_resolve_num = parisc_syscall_resolve_num,
+	.syscall_resolve_name_raw = parisc_syscall_resolve_name,
+	.syscall_resolve_num_raw = parisc_syscall_resolve_num,
 	.syscall_rewrite = NULL,
 	.rule_add = NULL,
 };
