+++ src/arch-parisc64.c
@@ -15,8 +15,8 @@ const struct arch_def arch_def_parisc64 = {
 	.token_bpf = AUDIT_ARCH_PARISC64,
 	.size = ARCH_SIZE_64,
 	.endian = ARCH_ENDIAN_BIG,
-	.syscall_resolve_name = parisc64_syscall_resolve_name,
-	.syscall_resolve_num = parisc64_syscall_resolve_num,
+	.syscall_resolve_name_raw = parisc64_syscall_resolve_name,
+	.syscall_resolve_num_raw = parisc64_syscall_resolve_num,
 	.syscall_rewrite = NULL,
 	.rule_add = NULL,
 };
